Add faster Gamma
